+++ piglegsurgeryweb/uploader/data_tools.py
@@ -1,5 +1,6 @@
 from pathlib import Path
 from typing import Optional, Union
+import json
 
 import gspread
 import pandas as pd
@@ -116,3 +117,47 @@ def remove_iterables_from_dict(dct: dict) -> dict:
     :return: dictionary without iterables
     """
     return {k: v for k, v in dct.items() if not hasattr(v, "__iter__")}
+
+
+class NumpyEncoder(json.JSONEncoder):
+    """ Special json encoder for numpy types """
+    def default(self, obj):
+        if isinstance(obj, np.integer):
+            return int(obj)
+        elif isinstance(obj, np.floating):
+            return float(obj)
+        elif isinstance(obj, np.ndarray):
+            return obj.tolist()
+        return json.JSONEncoder.default(self, obj)
+
+
+def save_json(data: dict, output_json: Union[str, Path], update: bool = True):
+    logger.debug(f"Writing '{output_json}'")
+
+    output_json = Path(output_json)
+    output_json.parent.mkdir(exist_ok=True, parents=True)
+    # os.makedirs(os.path.dirname(output_json), exist_ok=True)
+    dct = {}
+    if update and output_json.exists():
+        with open(output_json, "r") as output_file:
+            dct = json.load(output_file)
+        logger.debug(f"old keys: {list(dct.keys())}")
+    dct.update(data)
+    logger.debug(f"updated keys: {list(dct.keys())}")
+    with open(output_json, "w") as output_file:
+        try:
+            json.dump(dct, output_file, indent=4, cls=NumpyEncoder)
+        except Exception as e:
+            logger.error(f"Error writing json file {output_json}: {e}")
+            logger.error(f"Data: {dct}")
+            print_nested_dict_with_types(dct, 4)
+
+            raise e
+
+def print_nested_dict_with_types(d: dict, indent: int = 0):
+    for k, v in d.items():
+        if isinstance(v, dict):
+            logger.debug(f"{' ' * indent}{k}:")
+            print_nested_dict_with_types(v, indent + 2)
+        else:
+            logger.debug(f"{' ' * indent}{k}: {type(v)}")